fixes and changes

+++ src/main/java/com/aurd/Student/Model/Entity/NotesCommentModel.java
@@ -27,7 +27,7 @@ public class NotesCommentModel {
     private Timestamp added_on;
 
     @Column(name = "added_by",nullable = false)
-    private int added_by;
+    private Long added_by;
 
 
     @Column(name = "type",nullable = false)
@@ -100,13 +100,11 @@ public class NotesCommentModel {
         this.added_on = added_on;
     }
 
-    public int getAdded_by() {
+    public Long getAdded_by() {
         return added_by;
     }
 
-    public void setAdded_by(int added_by) {
+    public void setAdded_by(Long added_by) {
         this.added_by = added_by;
     }
-
-
 }
